feat: update booking flow and CMS content

+++ plugins/noren/booking/updates/builder_table_update_noren_booking_transfer_4.php
@@ -1,5 +1,4 @@
-<?php
-namespace Noren\Booking\Updates;
+<?php namespace Noren\Booking\Updates;
 
 use Schema;
 use October\Rain\Database\Updates\Migration;
@@ -8,14 +7,16 @@ class BuilderTableUpdateNorenBookingTransfer4 extends Migration
 {
     public function up()
     {
-        Schema::table('noren_booking_transfer', function ($table) {
-            $table->decimal('bus_price', 10, 2)->nullable();
+        Schema::table('noren_booking_transfer', function($table)
+        {
+            $table->integer('bus_price')->nullable();
         });
     }
-
+    
     public function down()
     {
-        Schema::table('noren_booking_transfer', function ($table) {
+        Schema::table('noren_booking_transfer', function($table)
+        {
             $table->dropColumn('bus_price');
         });
     }
